# visus-front/packages/api/openapi.json
{
    "openapi": "3.1.0",
    "info": { "title": "OpenAPI definition", "version": "v0" },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/gestor/{id}": {
            "get": {
                "tags": ["Módulo de Gestores"],
                "summary": "Obtém um Gestor de acordo com seu ID",
                "operationId": "getGestorById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gestor obtido com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GestorDetailsDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ocorreu erro durante o processamento da requsição",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GestorDetailsDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não foi possível obter o Gestor com o ID especificado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GestorDetailsDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["Módulo de Gestores"],
                "summary": "Atualiza as informações de um Gestor",
                "operationId": "updateGestor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGestorDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Gestor obtido com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GestorDetailsDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ocorreu erro durante o processamento da requsição",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GestorDetailsDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não foi possível obter o Gestor com o ID especificado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GestorDetailsDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/gestor": {
            "get": {
                "tags": ["Módulo de Gestores"],
                "summary": "Obtém um Gestor de acordo com seu e-mail",
                "operationId": "getGestorByEmail",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GestorGetByEmailDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Gestor obtido com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GestorDetailsDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ocorreu erro durante o processamento da requsição",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GestorDetailsDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não foi possível obter o Gestor com o ID especificado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GestorDetailsDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Endereco": {
                "type": "object",
                "properties": {
                    "logradouro": { "type": "string", "minLength": 1 },
                    "numero": { "type": "string", "minLength": 1 },
                    "bairro": { "type": "string", "minLength": 1 },
                    "cidade": { "type": "string", "minLength": 1 },
                    "estado": { "type": "string", "minLength": 1 },
                    "cep": { "type": "string", "minLength": 1 }
                },
                "required": [
                    "bairro",
                    "cep",
                    "cidade",
                    "estado",
                    "logradouro",
                    "numero"
                ]
            },
...
            "InstituicaoDetailsDTO": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer", "format": "int64" },
                    "nome": { "type": "string" },
                    "telefone": { "type": "string" },
                    "email": { "type": "string" },
                    "cnpj": { "type": "string" },
                    "endereco": { "$ref": "#/components/schemas/Endereco" }
                }
            },
            "UpdateGestorDTO": {
                "type": "object",
                "properties": {
                    "nome": { "type": "string" },
                    "telefone": { "type": "string" },
                    "genero": {
                        "type": "string",
                        "enum": [
                            "MASCULINO",
                            "FEMININO",
                            "OUTRO",
                            "NAO_INFORMAR"
                        ]
                    },
                    "dataNascimento": { "type": "string", "format": "date" },
                    "matricula": { "type": "string" },
                    "cargo": { "type": "string" }
                }
            },
            "GestorDetailsDTO": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer", "format": "int64" },
                    "nome": { "type": "string" },
                    "cpf": { "type": "string" },
                    "email": { "type": "string" },
                    "telefone": { "type": "string" },
                    "genero": {
                        "type": "string",
                        "enum": [
                            "MASCULINO",
                            "FEMININO",
                            "OUTRO",
                            "NAO_INFORMAR"
                        ]
                    },
                    "dataNascimento": { "type": "string", "format": "date" },
                    "cargo": { "type": "string" },
                    "instituicao": {
                        "$ref": "#/components/schemas/InstituicaoDetailsDTO"
                    }
                }
            },
...
            "GestorGetByEmailDTO": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    }
                },
                "required": ["email"]
            },
...
        }
    }
}